Swap GORM to Postgres with owned schema, hooks, and atomic upserts
Move OK Folio onto its own dedicated Postgres instead of the legacy MariaDB.

Driver/config:
- Swap gorm.io/driver/mysql -> gorm.io/driver/postgres (pgx stdlib); drop the
  MySQL driver, keep sqlite for unit tests.
- Rewrite DSN() to key/value form with TimeZone=UTC; DATABASE_URL returned
  verbatim. Add DB_PORT/DB_SSLMODE/DATABASE_URL env parsing and Port/SSLMode/URL
  on the config struct; defaults point at the postgres service on 5432, sslmode
  disable.
- Boot guard: refuse a legacy DB_HOST, assert the postgres dialect, and reject a
  legacy MySQL DSN before gorm.Open (testguard); CI lint blocks the mysql driver
  in production code.

Schema/models:
- uint64 ids; text/timestamptz/bytea mapping; PK promoted to GENERATED BY
  DEFAULT AS IDENTITY via idempotent post-migrate Exec so the loader can insert
  legacy ids.
- url_hash bytea NOT NULL (unique) populated by a single BeforeSave hook; raw url
  carries no btree. Re-key IsPhotoDownloaded/RecordDownload/MarkPhotoFailed/
  RecordFailedDownload to url_hash; drop url from the search predicate.
- Atomic upserts via clause.OnConflict; convert-loser-to-Inbox on 23505.
- Net-new columns: favorite, provider, category (derived, replaces the N+1
  filter), content_hash, perceptual_hash (bigint); embedding vector(512) added
  via guarded post-migrate Exec (no CREATE EXTENSION, never Fatal).
- (downloaded_at, artist) composite index; standalone downloaded_at index
  dropped. Native time scan replaces CAST AS CHAR; search uses ILIKE.
- Remove the favorite-column runtime probe; document 'pending' as transient.

Tests: Postgres integration test (real migrate + idempotency + ILIKE + composite
order + url_hash/content_hash concurrency) gated by OKFOLIO_TEST_POSTGRES_DSN and
a CI pgvector service container; sqlite kept for fast unit tests.

Refs #16

+++ internal/api/streams.go
@@ -43,7 +43,7 @@ type connectorSourceStatus struct {
 }
 
 type connectorRunStatus struct {
-	ID               uint       `json:"id"`
+	ID               uint64     `json:"id"`
 	StartTime        time.Time  `json:"start_time"`
 	EndTime          *time.Time `json:"end_time"`
 	Status           string     `json:"status"`
@@ -268,8 +268,8 @@ func maxTime(current *time.Time, candidate time.Time) *time.Time {
 	return current
 }
 
-func strconvUint(value uint) string {
-	return strconv.FormatUint(uint64(value), 10)
+func strconvUint(value uint64) string {
+	return strconv.FormatUint(value, 10)
 }
 
 func connectorProviderIDFromSource(sourcePage string, storedURL string) string {
